+++ docs/changelog.md
@@ -11,6 +11,7 @@
 + Add "auto remove empty links after deleted points" option. (github issue #23)
 + Move background option into project file. (github issue #26)
 + The background image can be set by dragging in.
++ Consider the background when zoom to fit.
 + Show the current mechanism in the overview dialog.  (github issue #25)
 + Support Solvespace 3.x format.  (github issue #28)
 + Replace term "workbook" into "project".